fix: speedup startup when getting version
