+++ antarest/gui.py
@@ -9,172 +9,45 @@
 # SPDX-License-Identifier: MPL-2.0
 #
 # This file is part of the Antares project.
-
-import contextlib
+import argparse
 import multiprocessing
-import platform
-import time
-import webbrowser
-from dataclasses import dataclass
-from multiprocessing import Process
-from pathlib import Path
-from threading import Thread
 
-import httpx
-import uvicorn
-from PyQt5.QtGui import QCursor, QIcon
-from PyQt5.QtWidgets import QApplication, QMenu, QSystemTrayIcon
-
-from antarest.core.utils.utils import get_local_path
-from antarest.main import fastapi_app, parse_arguments
-
-RESOURCE_PATH = get_local_path() / "resources"
+from antarest import __version__
+from antarest.core.cli import PathType
 
 
-def run_server(config_file: Path) -> None:
-    app = fastapi_app(
-        config_file,
-        mount_front=True,
-        auto_upgrade_db=True,
-    )[0]
-    # noinspection PyTypeChecker
-    uvicorn.run(app, host="127.0.0.1", port=8080)
-
-
-def start_server(config_file: Path) -> Process:
-    server = multiprocessing.Process(
-        target=run_server,
-        args=(config_file,),
+def parse_arguments() -> argparse.Namespace:
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "-c",
+        "--config",
+        type=PathType(exists=True, file_ok=True),
+        dest="config_file",
+        help="path to the config file [default: '%(default)s']",
+        default="./config.yaml",
     )
-    server.start()
-    return server
-
-
-def open_app() -> None:
-    webbrowser.open("http://localhost:8080")
-
-
-@dataclass(frozen=True)
-class AntaresSystrayApp:
-    """
-    Used to keep ownership of root Qt objects.
-    QMenu can only be owned by QWidgets, but we don't have one.
-    """
-
-    app: QApplication
-    menu: QMenu
-
-
-def create_systray_app() -> AntaresSystrayApp:
-    """
-    Creates the small application that allows to open
-    the browser or shutdown the server.
-    """
-    app = QApplication([])
-    app.setQuitOnLastWindowClosed(False)
-
-    # Adding an icon
-    icon = QIcon(str(RESOURCE_PATH / "webapp" / "logo16.png"))
-    # Adding item on the menu bar
-    tray = QSystemTrayIcon(icon, app)
-    tray.setToolTip("AntaresWebServer")
-
-    # Creating the options
-    menu = QMenu()
-    open_app_action = menu.addAction("Open application")
-    assert open_app_action is not None
-    open_app_action.triggered.connect(open_app)
-    # To quit the app
-    quit_action = menu.addAction("Quit")
-    assert quit_action is not None
-    quit_action.triggered.connect(app.quit)
-
-    # Adding options to the System Tray
-    def handle_action(reason: int) -> None:
-        """
-        Shows context menu also on left click
-        """
-        if reason == QSystemTrayIcon.Trigger:  # type: ignore
-            menu = tray.contextMenu()
-            assert menu is not None
-            menu.popup(QCursor.pos())
-
-    tray.setContextMenu(menu)
-    tray.activated.connect(handle_action)
-
-    tray.setVisible(True)
-
-    return AntaresSystrayApp(app, menu)
-
-
-def monitor_server_process(server: Process, app: QApplication) -> None:
-    """
-    Quits the application when server process ends.
-    """
-    server.join()
-    app.quit()
-
-
-def setup_exit_application_on_server_end(server: Process, app: QApplication) -> None:
-    Thread(target=monitor_server_process, args=(server, app)).start()
-
-
-def check_server_started() -> bool:
-    with contextlib.suppress(httpx.ConnectError):
-        res = httpx.get("http://localhost:8080/health")
-        if res.status_code == 200:
-            return True
-    return False
-
-
-def wait_for_server_start() -> None:
-    for _ in range(30, 0, -1):
-        if check_server_started():
-            break
-        time.sleep(1)
-
-
-def notification_popup(message: str, threaded: bool = True) -> None:
-    if platform.system() == "Windows":
-        # noinspection PyPackageRequirements
-        from win10toast import ToastNotifier  # type: ignore
-
-        toaster = ToastNotifier()
-        toaster.show_toast(
-            "AntaresWebServer",
-            message,
-            icon_path=RESOURCE_PATH / "webapp" / "favicon.ico",
-            threaded=threaded,
-        )
-    else:
-        from plyer import notification  # type: ignore
-
-        notification.notify(
-            title="AntaresWebServer",
-            message=message,
-            app_name="AntaresWebServer",
-            app_icon=str(RESOURCE_PATH / "webapp" / "favicon.ico"),
-            timeout=600,
-        )
+    parser.add_argument(
+        "-v",
+        "--version",
+        action="version",
+        help="Display the server version and exit",
+        version=__version__,
+    )
+    return parser.parse_args()
 
 
 def main() -> None:
+    print("Starting main execution...")
+
     multiprocessing.freeze_support()
 
     arguments = parse_arguments()
-    if check_server_started():
-        notification_popup(
-            "Antares Web Server already running, you can manage the application within the system tray.", threaded=False
-        )
-        return
-    notification_popup("Starting Antares Web Server...")
-    systray_app = create_systray_app()
-    server = start_server(arguments.config_file)
-    setup_exit_application_on_server_end(server, systray_app.app)
-    wait_for_server_start()
-    notification_popup("Antares Web Server started, you can manage the application within the system tray.")
-    systray_app.app.exec_()
-    server.kill()
+
+    # VERY important to keep this import here in order to have fast startup
+    # when only getting version
+    from antarest.desktop.systray_app import run_systray_app
+
+    run_systray_app(arguments.config_file)
 
 
 if __name__ == "__main__":
